Revert "refactor road_network_generator.py"
This reverts commit 7dafc5228d0a737bc0923ba27aabc3a593e6d499.
